update: split the css into three sections, remove ContactPage.tsx
