+++ components/general/SendMsg.tsx
@@ -1,4 +1,4 @@
-import styles from "@/styles/ContactPage.module.css";
+import styles from "@/styles/SendMsg.module.css";
 import { fontSyne, fontRobotoMono } from "@/pages/_app";
 import SocialMedia from "./SocialMedia/SocialMedia";
 import Button from "../base/Button";
