+++ pages/contact.tsx
@@ -1,15 +1,22 @@
 import Head from "next/head";
-import ContactPage from "@/components/general/ContactPage";
+import SendMsg from "@/components/general/SendMsg";
+import OurOffices from "@/components/general/OurOffices";
+import CommonQuestions from "@/components/general/CommonQuestions";
 
 export default function Contact() {
   return (
     <>
       <Head>
-        <title>Contact</title>
-        <meta name="description" content="Generated by Taodemy" />
+        <title>Contact page - TaoTech</title>
+        <meta name="description" content="Way to contact us. Find the common questions." />
+        <meta name="keywords" content="contact infomation, support, inquiry, help" />
+        <meta name="author" content="Taodemy Team" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
       </Head>
       <main>
-        <ContactPage />
+        <SendMsg />
+        <OurOffices />
+        <CommonQuestions />
       </main>
     </>
   );
