+++ styles/CommonQuestions.module.css
@@ -1,0 +1,33 @@
+.questions {
+  width: 100%;
+  height: 650px;
+  display: flex;
+  align-items: center;
+}
+.questions_title__contrainer {
+  width: 50%;
+  height: 80%;
+}
+.questions_right__container {
+  width: 50%;
+  padding-left: 50px;
+  border-left: 1px solid var(--border-default);
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+}
+.questions_right__question {
+  width: 100%;
+  padding-bottom: 40px;
+  margin-bottom: 40px;
+  border-bottom: 1px solid var(--border-default);
+}
+.questions_right__question:last-child {
+  border-bottom: none;
+  padding-bottom: 0;
+  margin-bottom: 0px;
+}
+.question__text {
+  margin-top: 10px;
+  color: var(--text);
+}
