--- styles/ContactPage.module.css
@@ -1,126 +1,0 @@
-.contact {
-  width: 100%;
-}
-.msg__container {
-  display: flex;
-  align-items: center;
-  height: 500px;
-  border-bottom: 1px solid var(--border-default);
-}
-.msg_left__container {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  padding-right: 10%;
-  width: 50%;
-  height: 80%;
-}
-.msg_left_title__container {
-  width: 100%;
-  height: 50%;
-}
-.msg_left_title__heading {
-  margin-bottom: 2%;
-}
-.social_media__container {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-.social {
-  margin-bottom: 2%;
-}
-.msg_form__container {
-  width: 35%;
-  height: 70%;
-}
-.label_input__container {
-  position: relative;
-}
-.msg__input {
-  width: 100%;
-  border-top: none;
-  border-left: none;
-  border-right: none;
-  border-width: 1px;
-  line-height: 280%;
-  background-color: var(--background-default);
-  margin-bottom: 5%;
-}
-.msg__textarea {
-  resize: none;
-}
-.msg__label {
-  position: absolute;
-  left: 0.2em;
-  line-height: 280%;
-  color: var(--text-default);
-}
-.offices {
-  width: 100%;
-  height: 450px;
-  display: flex;
-  align-items: center;
-  border-bottom: 1px solid var(--border-default);
-}
-.offices_left__container {
-  width: 33%;
-  height: 75%;
-}
-.offices_address__container {
-  width: 66%;
-  height: 75%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-.address_details__container {
-  width: 100%;
-  height: 35%;
-  display: flex;
-}
-.address_border__container {
-  width: 100%;
-  height: 30%;
-  padding-left: 35px;
-  border-left: 1px solid var(--border-default);
-  display: flex;
-  align-items: center;
-}
-.address__border_center {
-  width: 100%;
-  border-top: 1px solid var(--border-default);
-}
-.questions {
-  width: 100%;
-  height: 650px;
-  display: flex;
-  align-items: center;
-}
-.questions_title__contrainer {
-  width: 50%;
-  height: 80%;
-}
-.questions_right__container {
-  width: 50%;
-  padding-left: 50px;
-  border-left: 1px solid var(--border-default);
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-}
-.questions_right__question {
-  width: 100%;
-  padding-bottom: 40px;
-  margin-bottom: 40px;
-  border-bottom: 1px solid var(--border-default);
-}
-.questions_right__question:last-child {
-  border-bottom: none;
-  padding-bottom: 0;
-  margin-bottom: 0px;
-}
-.question__text {
-  margin-top: 10px;
-  color: var(--text);
-}
