+++ styles/OurOffices.module.css
@@ -1,0 +1,35 @@
+.offices {
+  width: 100%;
+  height: 450px;
+  display: flex;
+  align-items: center;
+  border-bottom: 1px solid var(--border-default);
+}
+.offices_left__container {
+  width: 33%;
+  height: 75%;
+}
+.offices_address__container {
+  width: 66%;
+  height: 75%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+.address_details__container {
+  width: 100%;
+  height: 35%;
+  display: flex;
+}
+.address_border__container {
+  width: 100%;
+  height: 30%;
+  padding-left: 35px;
+  border-left: 1px solid var(--border-default);
+  display: flex;
+  align-items: center;
+}
+.address__border_center {
+  width: 100%;
+  border-top: 1px solid var(--border-default);
+}
