+++ styles/SendMsg.module.css
@@ -1,0 +1,55 @@
+.msg__container {
+  display: flex;
+  align-items: center;
+  height: 500px;
+  border-bottom: 1px solid var(--border-default);
+}
+.msg_left__container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  padding-right: 10%;
+  width: 50%;
+  height: 80%;
+}
+.msg_left_title__container {
+  width: 100%;
+  height: 50%;
+}
+.msg_left_title__heading {
+  margin-bottom: 2%;
+}
+.social_media__container {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
+.social {
+  margin-bottom: 2%;
+}
+.msg_form__container {
+  width: 35%;
+  height: 70%;
+}
+.label_input__container {
+  position: relative;
+}
+.msg__input {
+  width: 100%;
+  border-top: none;
+  border-left: none;
+  border-right: none;
+  border-width: 1px;
+  line-height: 280%;
+  background-color: var(--background-default);
+  margin-bottom: 5%;
+}
+.msg__textarea {
+  resize: none;
+}
+.msg__label {
+  position: absolute;
+  left: 0.2em;
+  line-height: 280%;
+  color: var(--text-default);
+}
